+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file. See [conventional commits](https://www.conventionalcommits.org/) for commit guidelines.
 
 ---
+## [sbexec-v0.4.3](https://github.com/tyrchen/sbe/compare/sbexec-v0.4.2..sbexec-v0.4.3) - 2026-09-05
+
+### Bug Fixes
+
+- allow sandboxed processes to signal their own child workers during teardown
+
+---
 ## [sbexec-v0.4.2](https://github.com/tyrchen/sbe/compare/sbexec-v0.4.1..sbexec-v0.4.2) - 2026-09-05
 
 ### Miscellaneous Chores